chore: request #45643 express: 4.21.2 -> 5.1.0
https://github.com/expressjs/express/blob/v5.1.0/History.md

This allows us to bump body-parser to the latest version to resolve
CVE-2025-13466 / https://github.com/expressjs/body-parser/security/advisories/GHSA-wqch-xfxh-vrr4
which could lead to a DoS.

No functional changes.
Easiest way to test is to re-build your stack, re-build and install the
RPM on your instance an restart the tuleap-realtime service.

Change-Id: I4999cd9284907f3c517fa747a2c90a5a8c3d140e