refactor: Invert detection of BurningParrot pages
part of request #39759: Get Rid of Flaming Parrot in Trackers

How to test:
No functional change expected. Only the following pages should display
in FlamingParrot theme:
- Tracker Administration > Fields
- Tracker Administration > Notifications
- Tracker Administration > Delete artifacts > Confirmation page
- Edit Artifact page
- Create New Artifact page
- Report (homepage of a tracker)

Why?
It's easier to track the remaining work: when we finish converting one
of those pages, we can remove the condition in
isNotYetMigratedToBurningParrot(). By default, everything is in
BurningParrot, instead of listing all the pages.

Change-Id: I34a5ae4f7cf7778e0e6a79e56136567c3631ab2d