feat: Create a new approval table
Part of request #39756 Get rid of Flaming Parrot in Document

When document has not an approbal table, the "Create a new one" button
now opens a modal instead of redirecting to old ui.

This modal allow you to select reviewers from users and user groups (2
multi-select box (ilst-picker and lazybox)). When submitting the table
is created with status "disabled" and an event is sent to top vue
component to 'refresh' the displayed item/table.

Note: When history is empty, REST route to get all approval tables fail
in the VersionnedFactory because getRow returns false. So just a little
fix btw.

*Testing:*

- With a document without any table (you can delete them if needed), go
to the new approval table tab and click on "Create a new one"
=> The modal should open.

- Fill the reviewers you want, and submit
=> The table should be created. It appears in history and as the table
status is disabled you have the info alert about it. In adminstration
you should see the selected reviewers.

Change-Id: I42bbf009a6145bb753d3fa0c31f5a7220b3a77b2