chore: Make `move-artifact-action` library displayable in BurningParrot
Part of request #39759: Get Rid of Flaming Parrot in Trackers

The ultimate goal of this series of patches is to convert Artifact View
to BurningParrot. To do this, we must first convert the header that
contains the Action button and everything it contains.

This patch allows the 'Move Artifact' modal to be displayed in
BurningParrot.

Until Artifact View is completely in BurningParrot, the Move Artifact
modal must continue to be displayed for FlammingParrot pages.
The strategy here is to duplicate the files in the library and modify
those that need to be modified for BurningParrot.

Then, on a FP page, we continue to diplay the initial modal, and on a BP
page, we display the BurningParrot modal.

Once everything is in BurningParrot, we will juste have to delete the
files FlamingParrot files.

Note: To facilitate the review, files ending with `*FlamingParrot` or
`*-fp` are the exact copies of the old files. That way, it makes better
diff and you can focus your attention on the BurningParrot changes.

How to test:
-- On Artifact View, click on Action > Move this artifact
=> It should display the modal in FP
=> No functional change expected

-- Add `|| true` into `trackerPlugin.php:510`
This is needed to load a BurningParrot page (Do not panic, it's normal
that everything is broken).

-- Replace `index-fp.ts` by `index.ts` into `Artifact.php:607`
This allows the BP modal to be loaded instead of the FP one.

-- On the Artifact View, find and click on the `Move this artifact` link
=> It should display a beautiful modal in BurningParrot
=> No functional change expected

Change-Id: I01949edebb3447489f5805594a2a0cbd456e18ec