feat: Use specific properties of text fields
Text fields have 3 specific properties:
* rows
* cols
* default value

We use those properties so that the fields displayed in the admin interface
match the ones displayed in the artifact view.

BREAKING CHANGE: cols property is now ignored (in admin UI for now). We
don't see any reason to restrict the width of a free text field. The
rows are important because it gives the user an hint of the expected
content.
See https://tuleap.net/artidoc/1689

You can test it at: `/plugins/tracker/fields/<tracker-id>/`

Part of request #39759: Get Rid of Flaming Parrot in Trackers

Change-Id: I43a6d8efa786d7c8ca30e7a427540b46a24921d9