refactor: Lazy load approval table app
Part of request #39756 Get rid of Flaming Parrot in Document

With time passing and features migrated to the vue app, the approval
table part begin to grow fat.

As it is not a every second used part of the app, it can be lazy loaded
when routing to it. This way the main app is lighter

*No functional changes expected*

Document and approval table should work as before

Change-Id: I5c68fea54562f02db12eeac206dade3c7e21e591