fix: Don't display fork repositories form when no destination is available
part of request #39761 Get rid of Flaming Parrot in Git service

Why?

When the user is not member of the current project, they are not allowed
to fork a repository in it --> The radio button is disabled.

When the user is not admin of a project having the Git service enabled,
there is no "Fork in another project" radio button not project select
box.

In this particular case, there is only a disabled radio button "fork in
current project", and submitting the form will end up with a feedback
error "you are not allowed to view this page".

The proposition is to not display the form and display a dedicated
warning instead to explain why forking is not possible.

Additionally, I've added some styling on the disabled "fork in current
project" radio button to make it actually looking disabled (opacity +
not-allowed cursor). The same goes for the fork path form element when
"fork in another project" mode is selected.

How to test:
- When the user is not member of the current project AND is admin of a
  project with no git service enabled
--> No form
--> Warning explaining that there aren't any forkable repositories AND
there aren't any fork destinations available

- When the user is not member of the current project AND is admin of a
  Git project
--> The form is displayed (unless there aren't any forkable
repositories)
--> The "fork in current project" radio button is disabled and actually
looks disabled

Change-Id: Ic8dfa279585c3253fa125303b38220f18309ec26