feat: Add modal for table review
Part of request #39756 Get rid of Flaming Parrot in Document

Add a new modal allowing reviewers to do their review. In the reviewers
table the link is replaced by a mini button opening a modal.

In the modal some informations about the table is displayed and a link
to the reviewed version is available.

The review select label has a little ? with a popover explaining all
review states.

When submitting, as the does not exist yet, the 404 Not Found error is
displayed in the modal.

*Testing:*

Got to an available approval table. Click on the review button (probably
"Not yet"), the review modal should open.

You can use it, and when submitting the error "Not Found" is displayed

Change-Id: I2b7e8b7f26a6463d645fdd0afc10869116323afe