feat: Introduce REST route docman_items/{id}/approval_table/review
Part of request #39756 Get rid of Flaming Parrot in Document

This route update user review on an existing approval table.

It takes in the body the review state, the comment and whether use wants
notification (or not)

Note: I also fix the app not refreshing well the reviewer table. The
history table is refreshed, but strangely it needs a whole page refresh
to see the new table state.

*Testing:*

In document app, you should be able to review an approval table.

Change-Id: Ia49231e15e42b011d23e0e311e8a98bc7468a814