feat: Add approval table administration modal part 1
Part of request #39756 Get rid of Flaming Parrot in Document

The administration button is now on the top right of the panel.
When clicking on it, it opens a modal with 3 sections.

- Global settings (current patch)
- Notifications (not yet)
- Reviewers (not yet)

A warning alert the user that the modal is still under development with
a link to old ui

In this patch we add the global settings section -> approval requester,
table status and comment.

As each section will be of some size, I choose to split the modal into
some child component with v-model for their fields. It seems to work
pretty seamlessly.

The update and delete buttons call not yet implemented api route leading
to a "Not Found" error.

Note: Lazybox user autocompleter was only for multiple select, I changed
it a little to allow single user select.

*Testing:*

When a document has an approval table and you have write permissions,
you should see the Administration button.

In the modal you can edit global settings. If you clear the requester
select, the update button is disabled

Update and Delete buttons show an error when clicked

Change-Id: If1639f80db730b336236533a0fb5309d2a9e8d5f