feat: inject tracker structure instead of doing REST calls
Initial implementation of fields usage was relying on REST call in order
to get the structure of the tracker. However this route filter
unreadable field: if a field is not readable, then it won't appear in
the fields list.
This is problematic in the context of fields usage, where we want to
organize fields, even if they are not readable in the artifact view.

Updating the behavior of the REST API would introduce a breaking change.
We could pass an argument to the REST call in order to retrieve all
fields, but instead, we now inject the tracker structure directly during
the mount of application. This improves the load time.

No functional changes in the REST API.

You can test it at: `/plugins/tracker/fields/<tracker-id>/` with a field
without any READ permission: before this changeset the field is not
displayed at all in the new UI.

Part of request #39759: Get Rid of Flaming Parrot in Trackers

Change-Id: I4a9dea7cbb70925b07e5eacb73e66695f5033b3f