feat: Add approval table administration modal part 2
Part of request #39756 Get rid of Flaming Parrot in Document

When clicking on administration button, it opens a modal with 3 sections

- Global settings (done)
- Notifications (current patch)
- Reviewers (not yet)

A warning alert the user that the modal is still under development with
a link to old ui

In this patch we add the notifications section -> notification type,
send notification now.

A popover helper is available on notification type label (?) to present
the 2 available types

A button allow user to send a reminder to table reviewers. It calls the
api and as the route does not exists yet it only show an error.

*Testing:*

The notification type select should be set on current value. All
clickable buttons (except cancel) should display an error message

Change-Id: I28837dc3da68fbcd0efb4e9bd2d9de71d577e9be