fix: Folder approval table is broken
Part of request #39756 Get rid of Flaming Parrot in Document

Of all items which can have approval tables (file, embedded, wiki, link,
folder), only folders are not versionable.

In previous patchs all the approval table app assert that items are
versionable and has some logic around version attributes. But folder are
not and thus the app is broken for them (the current table not shown
because as version is null it is an exception case and the table is set
to null too -> no display)

Plus, as folders are not versionned, there is no meaning to display a
history of tables, there is only one.

*Testing:*

You can use approval table for folders. Other types of document are not
impacted

Change-Id: I89c2cb027a5bfbaaa5378cc8deb916073d6db87d