feat: Enforce basic guardrailing to be in place when doing completions with Mistral
This leverages the basic guardrailing offered by Mistral [0].
For the only current usage in cross-tracker search to create TQL queries
that seems good enough.

Burning more tokens and time to setup moderation or force the model to
do self-reflection [1] does not seem appropriate for this specific feature.

[0] https://docs.mistral.ai/capabilities/guardrailing#guardrailing
[1] https://evjang.com/2023/03/26/self-reflection.html

Part of request #45973 Adjust AI TQL Assistant with the secure coding guidelines

Change-Id: Ic0c8b2943a52029b94b0fb931c7072f135e22712