fix: Ensure TQL assistant endpoint is only accessible to authenticated users
This is an expensive endpoint, it should only accessible to
authenticated users.
The AI plugin has been reworked a bit to be sure we always pass
information about the requestor. For now this make sure we cannot call
it without being authenticated. Later on this information will also be
useful to collect metrics and give a bit of context in the logs.

Part of request #45973 Adjust AI TQL Assistant with the secure coding guidelines

Change-Id: I8faf505b0deb89549c871016119f2d6efbf57fac