feat: Add approval table administration modal part 3
Part of request #39756 Get rid of Flaming Parrot in Document

When clicking on administration button, it opens a modal with 3 sections

- Global settings (done)
- Notifications (done)
- Reviewers (current patch)

A warning alert the user that the modal is still under development with
a link to old ui

In this patch we add the reviewer section -> reviewer table, reminders
and add reviewer

The table present current reviewers and allow to change their rank, send
reminder and remove them. A info message tell the user that all these
changes are not saved and a click on the 'Update' button is *required*.

Below the table there is 2 inputs (the same as in table creation modal)
allowing user to choose some additional user and user group to add to
the table

*Testing:*

You should be able to use the table and the inputs.

When clicking on send reminder, you should get an error as API does not
exists yet.

Change-Id: I6684ae4fe55650fdbcccdc5dd841e9e26a873dd0