refactor: extract pull-request title into a custom element
part of request #44480 Kill AngularJS in Pullrequest plugin

We need a custom element to reuse between the three tabs
overview/commits/changes.

The role of this component is to display a skeleton while it's loading
the pull-request title, and to display it (+ tooltip) when done.

Note: The custom element is not used in the angular app because it is
planed to be removed once the changes tab is migrated to Vue3.

How to test:
- Go to the overview of a pull-request
--> The title is loaded
--> It has a tooltips if it contains cross-references

- Modify the title
--> The title has been updated (tooltips too)

- Go to the commits tab of the pull-request
- Same expectations

Change-Id: Ia23748bb96ac970a2f84a5301e405ddd56968901