feat: Add approval table administration modal - missing table
Part of request #39756 Get rid of Flaming Parrot in Document

Last part of administration modal.

Sometimes, versionned document with approval tables does not have a
table for their last version (it can be have deleted). In this case the
document has approval tables but none are usuable (the last table is
linked to an old version).

Old behavior in administration was letting user choose between 4
options:

- keep the current table (in fact it does nothing)
- copy the previous table
- reset the table
- create a new table

As the first option does nothing, I decided to keep only the 3
remaining.

When your document is in the described case, the modal shows only one
section with the 3 options. The update button then call a different API
(not implemented yet - 404).

*Testing:*

Create a file which can be versionned (file, embedded, link or wiki)
with at least 2 version each with an approval table.

In the old ui, delete the last table.

Go back to new ui, the admin modal should show you the options.

When clicking on update you should get a 404 error

Change-Id: I61a578a17119e7866f4d2e63c5952a725c6fe277