feat: Display date field with submitter timezone
Part of request #44843 Date fields should be displayed with submitter timezone

The whole feature is hidden behind a hidden feature flag:

    feature_flag_display_date_with_submitter_timezone

For each date fields in artifact view (datetime are not changed), the
date is displayed with changeset submitter timezone.

For that each time we call DateFormatter::fetchArtifactValue, the
changeset submitter timezone is retrieved and used in the TimezoneWrapper

*Testing:*

Enable the feature flag:

    tuleap config-set feature_flag_display_date_with_submitter_timezone 1

Have a user with the Europe/Paris timezone, submit an artifact with at
least one date field.

With another user with the America/Los_Angeles timezone, display the
artifact in the artifact view, the date field should display the date
according to Paris timezone.

If you reset the feature flag to 0, the date field should display the
date according to Los Angeles timezone.

Other display and datetime fields should not be impacted (no error 500)

Change-Id: I4565242bded2c7f51baeb6837b5b6d6040589a8b