chore: eslint: 8.57.0 -> 9.39.1
Implements request #42992 eslint: 8.52.0 -> 9.x

No functional change. CI should be happy. Configurations should still
flag code that breaks the rules in:
- JS tests
- TS tests
- Vue files
- Cypress tests

Why?
eslint re-introduced the "extends" and "globalIgnores" configurations
[0]. It lets us slightly simplify the configuration. As a result, Vue
rules are not applied to tests anymore, but it should be okay.

no-inner-declarations's behaviour has changed in eslint v9 [1], which is
why it is no longer reported.

[0]
https://eslint.org/blog/2025/03/flat-config-extends-define-config-global-ignores/
[1]
https://eslint.org/docs/latest/use/migrate-to-9.0.0#no-inner-declarations

Change-Id: Idc2013f514384f9cf631441d661ff3811e1319a2