feat: Collect metrics on the usage of LLMs
This should allow to detect abuse done by users.
Depending on the situation the number of users might not be finite
which could lead to a high cardinality metric. In this specific case
this is likely OK:
* handling the cost of a potentially high-cardinality metric will still
be lower than the cost associated with the usage of LLM APIs
* in real scenario without abuse the cardinality should be quite
  reasonable
* worst case scenario, organization should still have a way to drop the
  label/set limitations

Part of request #45973 Adjust AI TQL Assistant with the secure coding guidelines

Change-Id: Ia7954b8dbebfed48c36d88dfd6f7fd88081ef02d