feat: Introduce PUT docman_items/{id}/approval_table
Part of request #39756 Get rid of Flaming Parrot in Document

This new REST route allow user to update last approval table of document
items.

It takes some arguments in the body:
- owner -> user id of the table owner
- status -> enum string for table status
- comment -> the table description
- notification_type -> enum string for table notifications type
- reviewers -> list of user id to set as reviewers
- reviewers_group_to_add -> list of user group id to add to reviewers

For reviewers and reviewers_group_to_add there is a little notice to
precise: the first parameter list wanted users as reviewers, previous
users + additional, thus if you have users 101 and 102 and want to
remove 102 you send [101]. The order of the list is important as it used
for the reviewer rank (useful for sequential notification). The second
parameter adds members of corresponding user groups after all the
previously added users.

Note: I choose to change the method for the updating, from my point of
view it is better if the update uses PUT, and the copy/reset/create for
missing table uses PATCH.

*Testing:*

In administration, change some things (table status, notifications,
comment, reivewers, ...) and then click on update.

The request should success and the data is refreshed with the new
settings.

Change-Id: I493f716d0cc96887e5ac60fcef6f4876cc75d6b1