feat: convert Gerrit pane of Git repo settings to Burning Parrot
part of request #39761 Get rid of Flaming Parrot in Git service

How to test:
- No functional change expected in Git repository > Settings > Webhooks
pane. The modals should still open when clicking on the buttons.
- Either forge the code or set up a local Gerrit server.
- When there is at least one configured Gerrit server, in Git repository
> Settings, a "Gerrit" pane should appear.
- When you click on it, you should see a page that lets you migrate the
current Git repository to Gerrit.
    - When you select a server, some JS will show/hide some warnings.
    - There should be a warning if the Git repository was already
      migrated once to another server (Gerrit.php L170)
    - If the server does not have the Delete plugin (Gerrit.php L169),
      the warning appears but there should be no button.
- If the repository is in queue for the migration to Gerrit (it's a
System event), if you go to the "Gerrit" pane, you should see an alert
at the top of the page.
- If the migration fails (SystemEvent_GIT_GERRIT_MIGRATION.php L64), if
you go to the "Gerrit" pane, you should see an alert. The TLP relative
date in it should show the date of the migration.
- After migrating successfully a repository to Gerrit, if you go to the
"Gerrit" pane, you should see a form that lets you disconnect the
repository from the Gerrit server.
- If you submit the form, you should see a confirmation page for the
disconnect.

Why?
I chose to merge the "repository-admin-webhooks" package into the
"repository-admin" one. It is generally harder to split packages than it
is to merge, but the repository admin scripts are small and they don't
have many dependencies. It should be easy to find where the JS and SCSS
for those parts are.

I also leveraged the getJavascriptViteAssets() method in Pane that was
recently introduced, so that the Hooks pane will look like the others.

The script_locale were previously called in git.js, which is now removed
completely, along with the legacy styles.

Change-Id: I9eb1c3f1afef858f7b9e5bdd7b1f5a077e6df220