feat: Introduce POST docman_items/{id}/approval_table/reminder
Part of request #39756 Get rid of Flaming Parrot in Document

This new REST route allow user to send a mail to reviewers according to
approval table notification type.

- All at once -> send a mail to **all** reviewers
- Sequential -> send a mail to next reviewer in queue

If notification are disabled, you'll get an error 400 "No notification
sent".

*Testing:*

In the administration modal (with the notifications on all at once or
sequential), click on "Send a mail reminder to approver(s)".

According to your notification type, reviewers should receive a
notification. And a success message is displayed

Change-Id: Id2356f1ca4fd710690e588a2cc0323dfac81ea7e