fix: Wrong version_name in reviewer REST representation
Part of request #39756 Get rid of Flaming Parrot in Document

In `ItemApprovalTableReviewerRepresentation`, when building
version_name, we concatenate 'version ' with the version number. But if
reviewers has not done its review yet, the version number is null and
then version_name='version '.

We should check before if it is not null.

*Testing:*

In reviewers table:

- reviewers with review on 'Not yet', the version column should be
  empty.
- reviewers with review on something else, the version column should
  display something.

Change-Id: Ieb9a0e66dd39c77e67fa86e340660622c504c110