refacto: Move XTS AI assistant chat into a Vue component
No functional changes expected.

The current code does nto follow the secure coding guidelines as it
mixes the use of the template engine (lit) and direct DOM manipulations.
That makes the code hard to follow as everything is mixed together and
could potentially lead to injections. This is already visible as the
user prompts are treated as HTML content instead of plaintext.
The code also does not take care of only manipulating content within the
context of the widget. It might interact with other widgets on the page.

This contribution moves back everything into the Vue app. Using lit
would be fine if the template engine was leveraged to do all the
rendering and not only some parts of it.
In the present situation the widget already pulls Vue, bringing another
template engine does not seem useful and increase the size code that
need to be shipped to the end users.

Additionally, this change:
* dedupes the SVG AI icon and optimize it
* uses a proper form to handle user submission improving the
  accessibility
* prevents multiple submissions while waiting for an answer
* adds syntax highlighting of the TQL queries
* sets an appropriate title when updating the query

Part of request #45973 Adjust AI TQL Assistant with the secure coding guidelines

Change-Id: Ic7a6a4e25dc2569a82b7b8a3c4715fee33c9217c