feat: Introduce POST docman_items/{id}/approval_table/reminder/{reviewer_id}
Part of request #39756 Get rid of Flaming Parrot in Document

This new REST route allow user to force send a mail to a reviewer no
matter the notification type of the approval table.

*Testing:*

In the administration modal (with at least one reviewer), click on the
"Send reminder" button of a reviewer. Corresponding user should receive
an email and a success message is displayed

Change-Id: I17d8eb366f04ca7fbfca5bc8b512cede55eb0a67