refactor: Force the LLM to return a structured output
This should limit deviations from the expected content and
it makes things easier to work with.

In a attempt to limit the context size, we now only add the TQL queries
generated by the model into the context instead of the full response.
That seems to be sufficient as the model already knows about TQL.

Minimal change has been made to the front-end for now, another
contribution will refactor it to take advantage of the structured output
and limit the guesses.

Note that this change forces the model to answer with only 1 query.
I consider this to be an acceptable trade off, end users can continue to
chat with the AI assistant to adjust the query.

Part of request #45973 Adjust AI TQL Assistant with the secure coding guidelines

Change-Id: Ib0a7cb1630a6cf42c5710d732817bab15cf8aac9