chore: Switch the model used by the TQL AI assistant to Devstral 2
This is one of the latest model released by Mistral to manage codebase.
It has the same pricing than the generalist currently used but it seems
to be a bit faster and to handle the task quite well.
It is also currently free which is not bad during our testing phase.

https://docs.mistral.ai/models/devstral-2-25-12
https://mistral.ai/fr/news/devstral-2-vibe-cli

Part of request #45973 Adjust AI TQL Assistant with the secure coding guidelines

Change-Id: I06514bf2a9d206f218e47b0063508a43b4f37eb1