docs: request #45991 Document required radio buttons in TLP design system
How to test:
No functional change. Run the following command:
  pnpm --filter=tlp-doc run storybook

In the Forms section, there should be no visible change for TLP inputs,
checkboxes and selects. There is a new option "Mandatory" for radio
buttons, it only appears when "with form label" option is active
(because the asterisk icon needs it). When active, the form element has
the red asterisk and the radio buttons have the "required" attribute.

Why?
I changed the other types of inputs to use "nothing" from lit-html. It
should make it even more obvious that nothing is displayed in some
cases. I also removed some spaces between the label text and the icons.
The icon is supposed to have its spacing done by CSS, the space
character is not needed.

Change-Id: I43f05f70811af7f94b8d58eed59d4b6846cd0b0f