feat: store and load message exchange in DB
Major refactoring needed, I didn't find how to break it down.

The overall logic is changed, the front end no longer manage the
persistency of the messages, it's the backend, in database. Consequence,
the REST route is simplified to only send one message at the time with
management of a thread for follow-ups.

The database design is simple to keep the change as low as possible, we
will extends later one (with token informations for instance).

This contrib breaks the widget because for time reason I didn't have the
time to modify the front end as well.

Part of: request #45973 Adjust AI TQL Assistant with the secure coding guidelines

Change-Id: I895b45ad219c8a4617459ad8a82b6c6335615e1b