fix: request #45640 Artifact link field not displayed after transition error
I do not understand exactly how it works, but it seems that if the
div.tracker_formelement_read_and_edit is wrapping the field itself and
its hidden input, the artifact view script does not remove the node.

Note: There is still one problem which need some work on the links
component -> if you have added a link from the artifact, in the error
view the link is not shown. This is because the component does not take
care of the initial value and simply fetch artifact current links.

*Testing:*

Update an artifact without respecting transition rules, the artifact
links field should be displayed

Change-Id: If653bb58bbdc673f8db320eaac998e4424dc8ae8