refacto: Manage Gitolite configuration without SSH
The Gitolite configuration files are now written directly at the
expected place and we trigger Gitolite updates manually. This remove
the need of updating an intermediate Git repository, commit the content
and then push it to the repository on the gitolite side (which would
then activate some hooks...).

The initial setup of the Gitolite admin directory is kept at this stage,
it will be removed in a future contribution.

GitoliteConfWriter has been mostly rewritten in a way that makes the DB
the only source of truth instead of having to rely on both the DB and
the current state of the filesystem. The only thing that still rely at
this stage on the current state of the filesystem is getting the list of
projects. This should also be refactored at a later stage, in the
current state the situation is (at least) as good as it is was previously.
It also simplifies the logic as we do not need to do a bunch of string
replacements.

Gitolite updates are now trigerred at the end of an action instead of
the end of the system event queue processing. While this might burn a
bit more CPU/IO cycles (not even sure since we do not need to shell out
to Git anymore), it makes the process easier to understand and it puts
us in a direction where it would be possible to use async events instead
of a system event queue.

To test, do not forget to update `/etc/sudoers.d/tuleap_gitolite`. In
production this will be done by the packaging.

Part of request #45992 Manage Git repositories without relying on SSHd

Change-Id: I7622ec9ed538c3ff6e8a069df62587b5a679360f