refactor: extract as much logic as possible from Resource
As Ressources objects cannot be tested, I tried to extract the logic in
a dedicted class (ChatThreadManager) with injection from Resource
object.

The widget extraction from ID and verification logic was distributed
across the stack without a good reason so everything is now centralized
at the beginning of the process.

In order to achieve this I had to refactor how XTS widget is fetched in
UserIsAllowedToSeeWidgetChecker. With this refactoring the REST test
were failing because originally it was possible to have a widget not
attached to a dashboard in checkUserIsAllowedToSeeWidget. This is no
longer possible and test setup was adjusted to reflect that.

Part of: request #45973 Adjust AI TQL Assistant with the secure coding guidelines

Change-Id: Ic29925c6d3b90c0bc4625358e4ffa2fe7bcc7abd