feat: Introduce palette in field usage
This contribution is only here to have a concrete palette in order to be
able to discuss about the L&F later on.

The palette is non fonctional, and does not contain every form elements
possible (missing plugins' ones, like TTM steps). There is no grouping of
elements, and their icon is not a final choice.

DnD is not implemented, you cannot drop elements from the palette yet.

Part of request #39759: Get Rid of Flaming Parrot in Trackers

Change-Id: I9848ab2382c0d7632e220537951d851d706ca1f2