feat: Clean up last FlamingParrot things in Git
Closes request #39761 Get rid of Flaming Parrot in Git service

How to test:
No functional change expected.

Why?
I chose to rename the addJavascriptViteAsset() method to not depend
specifically on Vite. It is now typed with JavascriptAssetGeneric
instead. It's always Vite underneath, but using the interface feels more
future-proof.

All pages in Git plugin are BurningParrot, but I'm wary of modifying the
event handler in gitPlugin. I already broke it once 😅, so I guess it has
to stay until the event itself is removed.

Change-Id: If7f84a7a6bbd6121c2f573f984a3106dbcd3f189