feat: display select box field values
Part of request #39759: Get Rid of Flaming Parrot in Trackers

Note 1: Only the values are displayed in this contribution. The decorators
will be implemented in a dedicated patch.

Note 2: The none value will be added in a dedicated patch too.

You can test it at: `/plugins/tracker/fields/<tracker-id>/`

How to test:
- Have three multiple selects bound to static, users and user groups
  values
- Have three single selects bound to static, users and user groups
- Define default values for these fields
--> The values are displayed as expected
--> The default values are selected

Change-Id: Ief67e220a78a10e68ce823f9ece0c3ee039c9d0f