refacto: Merge PR without opening a SSH connection
No functional changes, it should still be possible to
merge pull-requests.
The only change this contribution brings is that it does not open an SSH
connection to do it.

To test, do not forget to update `/etc/sudoers.d/tuleap_plugin_pull_request`.
This is managed by the packaging in production.

Part of request #45992 Manage Git repositories without relying on SSHd

Change-Id: I6758372b0d8eeb3cbfc400239ece5a3e5d1ef7b3