fix: request #46002 Missing index on the submitted_on key of the tracker_changeset table
This column is used when building the burnup and volocity graphs and can
also be used by some TQL queries. The DB should have a shortcut to be
able to quickly filter data.

Outside the scope for this change but the burnup query would benefit to
be rewritten using some CTEs. It would reduce the amount of data needing
to be traversed.

Change-Id: I502a3c27b7d8dfee808ef1d984092735979e16f5