chore: request #46327 Update dev/build environment [W1 2026]
Main changes:
* Go: 1.25.3 -> 1.25.5
* Composer: 2.8.12 -> 2.9.2
* CDXGen: 11.0.3 -> 11.10.0
* OSV Scanner: 2.2.3 -> 2.3.0
* PHP: 8.4.13 -> 8.4.16
* NodeJS: 24.10.0 -> 24.12.0
* Trufflehog: 3.90.9 -> 3.92.4
* Zizmor: 1.15.2 -> 1.19.0
* Docker Compose: 2.39.4 -> 5.0.0
* Oxipng: 9.1.5 -> 10.0.0
* coreutils: 9.7 -> 9.8
* Git: 2.51.0 -> 2.52.0

I have kept editorconfig-checker to an old versions.
The recent versions has an issue wit the charset detection with some of
our files. Upstream is aware of the issue.

Also kept the old ref to build the Git binary used by Tuleap in
production to not have to deal with patch adjustments. Upgrade to Git
2.52 for production usages will be handled in a dedicated change.

Change-Id: Ia7df8198f43d8a3a6039332489391be506497858