fix: Avoid crash when cleaning temporary PR merge directory containing a broken symlink
After the merge of a PR, cleaning up the temporary directory that was
used to handle the merge should not throw if the directory contains a
broken symlink.

Note this should also prevent the issue that are sometimes visible
while cleaning the Valinor cache during REST tests init.

Part of request #45992 Manage Git repositories without relying on SSHd

Change-Id: Ieea7912bc9252af221c31e71a121f7df2a0fe7ac